a new version of menu
this still needs work. you can register, and then sign in but it will only write a message to console saying hello (name).
Just a better menu. also i made some random drivers with location. they are in the list everytime we run the program. so if some one can connect them to the logic using the menu that will be great.